# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Media Tech BMSTU Equipment
  description: |
    REST API для автоматизации бронирования оборудования и Студии 626 Media Tech BMSTU.
    
    ## Роли пользователей
    - **Trainee** (Стажёр) — базовый доступ, ограниченный рейтингом "Порядочность"
    - **Activist** (Активист) — полный доступ к бронированию
    - **Admin** (Администратор) — управление системой, может стать куратором заявок
    
  version: 0.0.1
  contact:
    name: Media Tech BMSTU Team
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Получить текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    patch:
      tags: [Users]
      summary: Обновить профиль текущего пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Профиль обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/me/invocations:
    get:
      tags: [Users]
      summary: Получить все бронирования текущего пользователя
      description: Объединённый список заявок на оборудование и студию
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [created, under_review, changes_required, approved, equipment_issued, equipment_returned, completed, cancelled]
            description: Фильтр по статусу заявки
        - name: type
          in: query
          schema:
            type: string
            enum: [equipment, studio, all]
            default: all
        - name: active_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Список бронирований
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment_invocations:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipmentInvocationResponse'
                  studio_invocations:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudioInvocationResponse'
  /users:
    get:
      tags: [Users]
      summary: Список всех пользователей (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: full_name
          in: query
          schema:
            type: string
          description: Фильтр по полному имени (частичное совпадение)
          example: "Иван"
          required: false
        - name: email
          in: query
          schema:
            type: string
          description: Фильтр по email (частичное совпадение)
          example: "user@example.com"
          required: false
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
        '403':
          description: Доступ запрещён
  /users/{id}:
    get:
      tags: [Users]
      summary: Получить пользователя по ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '403':
          description: Доступ запрещён
        '404':
          description: Пользователь не найден

    patch:
      tags: [Users]
      summary: Обновить пользователя (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAdminDto'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Невалидные данные или конфликт
        '403':
          description: Доступ запрещён
        '404':
          description: Пользователь не найден
        '409':
          description: Email уже используется

  # ==================== ORGANIZATIONS ====================
...
  /organizations/{id}/users:
    get:
      tags: [Organizations]
      summary: Получить пользователей организации
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

  # ==================== DEPARTMENTS ====================
...
  /departments/{id}/users:
    get:
      tags: [Departments]
      summary: Получить пользователей отдела с ролями
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/RoleInDepartment'
      responses:
        '200':
          description: Список пользователей с ролями
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    user:
                      $ref: '#/components/schemas/UserResponse'
                    role:
                      $ref: '#/components/schemas/RoleInDepartment'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleInDepartment:
      type: string
      enum: [trainee, activist]
      description: Роль пользователя в отделе
...
    EquipmentInvocationStatus:
      type: string
      enum: [created, under_review, changes_required, approved, equipment_issued, equipment_returned, completed, cancelled]
      description: |
        Статусы заявки на оборудование:
        - created — Создана
        - under_review — На рассмотрении
        - changes_required — Требуются изменения
        - approved — Одобрена
        - equipment_issued — Выдано оборудование
        - equipment_returned — Сдано оборудование
        - completed — Завершена
        - cancelled — Отменена
    StudioInvocationStatus:
      type: string
      enum: [created, under_review, changes_required, approved, completed, cancelled]
      description: |
        Статусы заявки на студию:
        - created — Создана
        - under_review — На рассмотрении
        - changes_required — Требуются изменения
        - approved — Одобрена
        - completed — Завершена
        - cancelled — Отменена

    # ==================== DTOs ====================
...
    UpdateUserDto:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
    UpdateUserAdminDto:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 255
          description: Новое ФИО пользователя
        email:
          type: string
          format: email
          description: Новый email (должен быть уникальным)
        password:
          type: string
          minLength: 8
          description: Новый пароль
        nice:
          type: integer
          minimum: 1
          description: Баллы порядочности (CHECK nice > 0)
        is_admin:
          type: boolean
          description: Назначить/снять права администратора
        organizations:
          type: array
          description: Полный список организаций пользователя (заменяет существующие)
          items:
            type: object
            required: [organization_id, action]
            properties:
              organization_id:
                type: string
                format: uuid
              action:
                type: string
                enum: [add, remove]
                description: add - добавить, remove - удалить
        departments:
          type: array
          description: Полный список отделов и ролей пользователя (заменяет существующие)
          items:
            type: object
            required: [department_id, role, action]
            properties:
              department_id:
                type: string
                format: uuid
              role:
                $ref: '#/components/schemas/RoleInDepartment'
              action:
                type: string
                enum: [add, update, remove]
                description: add - добавить, update - изменить роль, remove - удалить
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
        nice:
          type: integer
          minimum: 1
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserDetailResponse:
      allOf:
        - $ref: '#/components/schemas/UserResponse'
        - type: object
          properties:
            organizations:
              type: array
              items:
                $ref: '#/components/schemas/OrganizationResponse'
            departments:
              type: array
              items:
                type: object
                properties:
                  department:
                    $ref: '#/components/schemas/DepartmentResponse'
                  role:
                    $ref: '#/components/schemas/RoleInDepartment'
            recent_invocations:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  type:
                    type: string
                    enum: [equipment, studio]
                  event_name:
                    type: string
                  status:
                    type: string
                  created_at:
                    type: string
                    format: date-time
...
    OrganizationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
...
    DepartmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
...
    EquipmentInvocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_name:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        equipment_return_time:
          type: string
          format: date-time
        sd_card_return_time:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/EquipmentInvocationStatus'
        curator_comment:
          type: string
          nullable: true
        admin_id:
          type: string
          format: uuid
          nullable: true
        admin_name:
          type: string
          nullable: true
        organization_id:
          type: string
          format: uuid
          nullable: true
        organization_name:
          type: string
          nullable: true
        department_id:
          type: string
          format: uuid
          nullable: true
        department_name:
          type: string
          nullable: true
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        equipment_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    StudioInvocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_name:
          type: string
        shooting_description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        needs_chromakey:
          type: boolean
        needs_cyclorama:
          type: boolean
        needs_black_fabric:
          type: boolean
        status:
          $ref: '#/components/schemas/StudioInvocationStatus'
        curator_comment:
          type: string
          nullable: true
        admin_id:
          type: string
          format: uuid
          nullable: true
        admin_name:
          type: string
          nullable: true
        organization_id:
          type: string
          format: uuid
        organization_name:
          type: string
        department_id:
          type: string
          format: uuid
        department_name:
          type: string
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        equipment_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...